+++ src/StitchPreview.ts
@@ -1,4 +1,6 @@
 import * as vscode from 'vscode';
+import * as path from 'path';
+import { Disposable } from './utils/dispose';
 import { CONSTANTS } from './constants';
 import { CommandAction, Context, ICommand } from './types';
 import { PdfPreview } from './PdfPreview';
@@ -10,84 +12,75 @@ import { EditorSimulateIntegrationResponse, ErrorData, StitchError } from './typ
 import { RenderTemplateStepResult } from './types/stepResult';
 import { HttpMulipartStepConfiguration, HttpStepConfiguration } from './types/stepConfiguration';
 
-export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposable {
+export class StitchPreview extends Disposable implements vscode.Disposable {
     
-    private _view?: vscode.WebviewView;
     private _result?: EditorSimulateIntegrationResponse;
     private _scrollPosition?: number;
     private _htmlHelper?: StitchPreviewHtmlBuilder;
-    private _disposables: vscode.Disposable[] = [];
 
-    public constructor(
-        private readonly _extensionUri: vscode.Uri
-    ) {}
+    public static create(extensionUri: vscode.Uri): StitchPreview {
 
-    public resolveWebviewView(
-        webviewView: vscode.WebviewView,
-        _context: vscode.WebviewViewResolveContext,
-        _token: vscode.CancellationToken,
-    ): void | Thenable<void> {
-        this._view = webviewView;
-
-        ContextHandler.log(`Stitch Preview webview resolved`);
-
-        webviewView.webview.options = {
-            enableScripts: true,
-            localResourceRoots: [vscode.Uri.joinPath(this._extensionUri, 'assets')]
+        const showOptions = {
+            viewColumn: vscode.ViewColumn.Two,
+            preserveFocus: true
         };
+        const options = {
+            enableScripts: true,
+            localResourceRoots: [vscode.Uri.file(path.join(extensionUri.path, 'assets'))]
+        };
+        const panel = vscode.window.createWebviewPanel('stitchPreview', 'Stitch Preview', showOptions, options);
+
+        panel.iconPath = vscode.Uri.joinPath(extensionUri, 'assets/icon.png');
+
+        return new StitchPreview(panel, extensionUri);
+    }
+
+    private constructor(
+        private _panel: vscode.WebviewPanel,
+        extensionUri: vscode.Uri
+    ) {
+        super();
+
+        this._panel.webview.html = '<html><body><h3>Loading preview...</h3></body></html>';
 
         const resolveAsUri = (...p: string[]): vscode.Uri => {
-            return webviewView.webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, ...p));
+            return this._panel.webview.asWebviewUri(vscode.Uri.joinPath(extensionUri, ...p));
         };
-        this._htmlHelper = new StitchPreviewHtmlBuilder(webviewView.webview.cspSource, resolveAsUri);
+        this._htmlHelper = new StitchPreviewHtmlBuilder(_panel.webview.cspSource, resolveAsUri);
 
-        webviewView.webview.onDidReceiveMessage(
-            (command: ICommand) => { ContextHandler.handlePreviewCommand(command, this._extensionUri); },
-            undefined,
-            this._disposables
+
+        const onPanelDisposeListener = this._panel.onDidDispose(() => this.dispose());
+        const onDidReceiveMessageListener = this._panel.webview.onDidReceiveMessage(
+            (command: ICommand) => { ContextHandler.handlePreviewCommand(command, extensionUri); }
         );
 
-        webviewView.onDidChangeVisibility(
-            () => {
-                ContextHandler.log(`Stitch Preview visibility changed: ${webviewView.visible}`);
-                if (webviewView.visible) {
-                    ContextHandler.showPreview(this);
-                } else {
-                    ContextHandler.hidePreview();
-                }
-            },
-            undefined,
-            this._disposables
-        );
-
-        webviewView.onDidDispose(
-            () => this.dispose(),
-            undefined,
-            this._disposables
-        );
-
-        // Request initial update if context is available
-        ContextHandler.showPreview(this);
+        this._register(this._panel);
+        this._register(onPanelDisposeListener);
+        this._register(onDidReceiveMessageListener);
     }
 
     public dispose(): void {
         PdfPreview.disposeAll();
-        while (this._disposables.length) {
-            const disposable = this._disposables.pop();
-            if (disposable) {
-                disposable.dispose();
-            }
-        }
-    }    
+        super.dispose();
+    }
+
+    public get visible(): boolean {
+        return this._panel?.visible ?? false;
+    }
+
+    public reveal(): void {
+        this._panel?.reveal(vscode.ViewColumn.Two);
+    }
 
     public handleCommand(command: ICommand, extensionUri: vscode.Uri): void {
         const response = this._result;
-        if (!response) { return; }
 
         ContextHandler.log(`Handling preview command: ${CommandAction[command.action]}, content: ${command.content}`);
 
         switch (command.action) {
             case CommandAction.viewStepRequest: {
+                if (!response) { return; }
+
                 const step = command.content;
                 const template = response.stepConfigurations[step].template?.trim();
                 if (!template) {
@@ -101,6 +94,8 @@ export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposa
                 return;
             }
             case CommandAction.viewStepResponse: {
+                if (!response) { return; }
+
                 if (response.integrationContext.steps[command.content]?.$type !== CONSTANTS.renderTemplateStepResultType) { return; }
                 const renderResponse = <RenderTemplateStepResult>response.integrationContext.steps[command.content];
                 if (renderResponse.response.contentType !== 'application/pdf') { return; }
@@ -109,6 +104,7 @@ export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposa
                 return;
             }
             case CommandAction.viewIntegrationResponse: {
+                if (!response) { return; }
                 StitchPreviewHelper.show({
                     filename: `stitch-response`,
                     content: unescapeResponseBody(response.result)
@@ -120,6 +116,7 @@ export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposa
                 return;
             }
             case CommandAction.createHttpRequest: {
+                if (!response) { return; }
                 const step = command.content;
                 const httpConfig = <HttpStepConfiguration>response.stepConfigurations[step];
                 StitchPreviewHelper.show({
@@ -129,6 +126,7 @@ export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposa
                 return;
             }
             case CommandAction.createHttpMultipartRequest: {
+                if (!response) { return; }  
                 const step = command.content;
                 const httpConfig = <HttpMulipartStepConfiguration>response.stepConfigurations[step];
                 StitchPreviewHelper.show({
@@ -146,16 +144,20 @@ export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposa
                 void vscode.window.showTextDocument(fileUri, { viewColumn: vscode.ViewColumn.One, preserveFocus: false });
                 return;
             }
+            case CommandAction.togglePinned: {
+                void ContextHandler.togglePinned();
+                return;
+            }
         }
     }
 
     public showSimulationResult(simulationResult: EditorSimulateIntegrationResponse, context: Context) {
-        if (!this._view || !this._htmlHelper) { return; }
+        if (!this._panel || !this._htmlHelper) { return; }
         
         this._result = simulationResult;
-        this._view.webview.html = this._htmlHelper.createHtml(context, simulationResult);
+        this._panel.webview.html = this._htmlHelper.createHtml(context, simulationResult);
         if (this._scrollPosition) {
-            void this._view.webview.postMessage({ command: 'setScrollPosition', scrollY: this._scrollPosition });
+            void this._panel.webview.postMessage({ command: 'setScrollPosition', scrollY: this._scrollPosition });
             this._scrollPosition = undefined;
         }
 
@@ -163,11 +165,11 @@ export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposa
     }
 
     public showErrorResult(errorData: ErrorData, context: Context) {
-        if (!this._view) { return; }
+        if (!this._panel) { return; }
         
         this._result = undefined;
         if (!this._scrollPosition) {
-            void this._view.webview.postMessage({ command: 'requestScrollPosition' });
+            void this._panel.webview.postMessage({ command: 'requestScrollPosition' });
         }
 
         if (errorData.ClassName === 'Core.Exceptions.StitchResponseSerializationException') {
@@ -184,21 +186,20 @@ export class StitchPreview implements vscode.WebviewViewProvider, vscode.Disposa
         }
     }
 
-
     public handleError(error: Error, title: string, context: Context) {
-        if (!this._view || !this._htmlHelper) { return; }
+        if (!this._panel || !this._htmlHelper) { return; }
         
         this._result = undefined;
-        this._view.webview.html = this._htmlHelper.createErrorHtml(context, {
+        this._panel.webview.html = this._htmlHelper.createErrorHtml(context, {
             title,
             description: error.message,
         }, JSON.stringify(error));
     }
 
     private _handleStitchError(context: Context, error: StitchError, extraBody?: string) {
-        if (!this._view || !this._htmlHelper) { return; }
+        if (!this._panel || !this._htmlHelper) { return; }
         
-        this._view.webview.html = this._htmlHelper.createErrorHtml(context, error, extraBody);
+        this._panel.webview.html = this._htmlHelper.createErrorHtml(context, error, extraBody);
     }
 }